Update README and add Page Views
I just updated the `README` to be a markdown and then added Page Views. 

about Page Views
When we add this service, it will count every hit of this repo. And this will guide us more about the visitors each day and will indicate the total views. For me, this is very helpful both for us and for those will view this repo seeing this page views. If there is the website built from this repo, it can be simply added there too.